backend/requirements.txt

django>=5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=23.5
psycopg2-binary>=2.9
Pillow>=10.0
django-unfold>=0.30.0
xhtml2pdf>=0.2.11
